requirements.txt
asgiref==3.5.2
Django==4.1
gobject==0.1.0
proxy-tools==0.1.0
pycairo==1.21.0
PyGObject==3.42.2
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pywebview==3.6.3
sqlparse==0.4.2
